Initial commit, with working verison of the tree app.
This is the initial verion of the tree app. It works, but not sure the
quality of the program. Still learning Go. I tried to use go routines,
but have run into two issues:

   1. I was running out of system threads.
   2. I was hitting the limit on the number of files open at one time.

To combat this I have a semphore that controls the creation of go
routines; and also use that to slow down the number of files we keep
open. It's really the number of directories that is being read. As we
don't open any files that are not directores. I think the code is pretty
simple, but I'm sure there are ways to improve it.

TODO.md contains a list of todos.